Release 0.2.0
- Worker refactor. The Worker behaviour has moved to `Lapin.Connection`
and the `Lapin.Pattern` behaviour is now decoupled from it and is per
channel. This is a backwards incompatible change, check out the docs for
the new configuration format. Sorry.

- Introduce `Lapin.Message.Payload` to encode/decode custom data types
automatically on publishing and consuming messages. This is backwards
compatible as the default implementation leaves binary unaltered. Again,
check the docs for implementation details.