Log stacktrace when rejecting a message because of an exception
Before this change only the top level error message would be shown. We now make sure to print the whole stacktrace of the error, which makes debugging an exception much easier.

The approach seems to be the go to solution as it's mentioned in the offficial docs: https://elixir-lang.org/getting-started/try-catch-and-rescue.html#reraise